requirements.txt
# Core framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
gunicorn==22.0.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.9.0
pydantic-settings==2.6.0

# Caching
redis[hiredis]==5.2.0

# AI/ML services
google-generativeai==0.8.0
anthropic==0.40.0
torch==2.5.0
torchvision==0.20.0
Pillow==11.0.0

# Google Cloud Platform
google-cloud-storage==2.18.0
google-cloud-secret-manager==2.21.0
cloud-sql-python-connector[asyncpg]==1.12.0

# Security & encryption
cryptography==43.0.0

# Utilities
python-dotenv==1.0.1
structlog==24.4.0
tenacity==9.0.0
jsonrepair==0.1.0
httpx==0.28.0
python-multipart==0.0.12
